+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * Add `$force` parameter to the `cache()` function to force recomputing the value
 * Add `getAttributes()` method to `TaskCommand` to get all or specific attributes
+* Add `CASTOR_MEMORY_LIMIT` environment variable to set the memory limit when 
+  using the `run_php()` function
 
 ### Fixes